Move `partial_path` to ObjectRendering
Partial paths are only required for deriving template info for rendering
objects, which is only the case for the object renderer and the
collection renderer
